Move out tile_element_height, coordinate_3d_to_2d.
