+++ src/openloco/map/tile.h
@@ -62,6 +62,15 @@ namespace openloco::map
     };
 #pragma pack(pop)
 
+    // TODO: rename struct
+    typedef struct LocationXY16
+    {
+        int16_t x, y;
+    } LocationXY16;
+
+    uint32_t tile_element_height(int16_t x, int16_t y);
+    LocationXY16 coordinate_3d_to_2d(int16_t x, int16_t y, int16_t z, int rotation);
+
     enum class element_type
     {
         surface,      // 0x00
@@ -248,4 +257,5 @@ namespace openloco::map
         size_t index_of(const tile_element_base* element) const;
         surface_element* surface();
     };
+
 }
